+++ src/components/EntityMetadata/styles.ts
@@ -1,0 +1,8 @@
+import { createStyles } from 'antd-style'
+
+export const useStyles = createStyles(() => ({
+  content: {
+    maxWidth: 340,
+    minWidth: 220
+  }
+}))